[trello.com/c/4QaDmVkg]: review fixes
