+++ Adamant/Modules/Wallets/Bitcoin/BitcoinKitTransactionFactoryProtocol.swift
@@ -19,25 +19,3 @@ protocol BitcoinKitTransactionFactoryProtocol {
         keys: [PrivateKey]
     ) -> Transaction
 }
-
-final class BitcoinKitTransactionFactory: BitcoinKitTransactionFactoryProtocol {
-    func createTransaction(
-        toAddress address: Address,
-        amount: UInt64,
-        fee: UInt64,
-        changeAddress: Address,
-        utxos: [UnspentTransaction],
-        lockTime: UInt32,
-        keys: [PrivateKey]
-    ) -> Transaction {
-        BitcoinKit.Transaction.createNewTransaction(
-            toAddress: address,
-            amount: amount,
-            fee: fee,
-            changeAddress: changeAddress,
-            utxos: utxos,
-            lockTime: lockTime,
-            keys: keys
-        )
-    }
-}
